modernize: apply clang-tidy modernization, safety, and configuration fixes

+++ src/buffer_pool.cpp
@@ -333,7 +333,8 @@ void BufferPool::trim() {
   const auto used = totalRows - trailing;
 
   const std::uint64_t wanted = std::max<std::uint64_t>(
-      static_cast<std::uint64_t>(used) + (used / 16 * trimHeadroom),
+      static_cast<std::uint64_t>(used) +
+          (static_cast<std::uint64_t>(used / 16) * trimHeadroom),
       trimFloorRows);
   if (wanted >= totalRows) {
     return;
